BackEnd/App/requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
google-auth==2.40.3
requests==2.32.4
python-dotenv==1.0.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
redis==5.0.1
gunicorn==21.2.0
uvicorn==0.27.1 